Weekly email script working
