+++ backend/pyproject.toml
@@ -38,6 +38,7 @@ geopy = "^2.4.1"
 ipykernel = "^6.29.4"
 install = "^1.3.5"
 pytest-django = "^4.8.0"
+django-ses = "^4.0.0"
 
 
 [build-system]
